+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  - [new] Update network traffic analysis settings
  - [new] List and create network webhook servers
  - [new] List, create, delete network webhook payload templates
+ - [new] Claim an order and add devices to network
 
 ### v0.1.6
  - [new] Update network settings